Part 2 of commit below (did not select all files initially).
Fixed the wanderer (it was still problematic).
Wanderer now moves in 3D space.

Added a bit of code to the Controller that gives the camera a standard starting position with a nice overview of the situation.
It now adds wanderer agents instead of red/blue functions for demonstration purposes. Simply comment / uncomment the respective lines in the controller to change it back.